+++ packages/react-core/src/core/hooks/wallet-hooks.ts
@@ -12,6 +12,7 @@ import type {
   MagicLink,
   MetaMaskWallet,
   OKXWallet,
+  OneKeyWallet,
   PaperWallet,
   PhantomWallet,
   RabbyWallet,
@@ -47,6 +48,7 @@ type WalletIdToWalletTypeMap = {
   phantom: PhantomWallet;
   walletConnectV1: WalletConnect;
   okx: OKXWallet;
+  oneKey: OneKeyWallet;
   cryptoDefiWallet: CryptoDefiWallet;
   rabby: RabbyWallet;
 };