+++ packages/react-core/src/evm/index.ts
@@ -219,32 +219,30 @@ export {
 // thirdweb hooks (work as long as at least `<ThirdwebSdkProvider>` is used)
 
 // auth hooks
-export {
-  useLogin,
-  useLogout,
-  useUser,
-  useAuth,
-  useSwitchAccount,
-} from "./hooks/auth";
-export type { UserWithData } from "./hooks/auth";
+export { useAuth } from "./hooks/auth/useAuth";
+export { useLogin } from "./hooks/auth/useLogin";
+export { useLogout } from "./hooks/auth/useLogout";
+export { useUser } from "./hooks/auth/useUser";
+export { useSwitchAccount } from "./hooks/auth/useSwitchAccount";
+
+export type { UserWithData } from "./hooks/auth/useUser";
 
 // contract hooks
-export {
-  useEditionDrop,
-  useEdition,
-  useNFTDrop,
-  useMarketplace,
-  useNFTCollection,
-  usePack,
-  useToken,
-  useTokenDrop,
-  useVote,
-  useSplit,
-  useMultiwrap,
-  useSignatureDrop,
-} from "./hooks/contracts";
+export { useEditionDrop } from "./hooks/contracts/useEditionDrop";
+export { useEdition } from "./hooks/contracts/useEdition";
+export { useNFTDrop } from "./hooks/contracts/useNFTDrop";
+export { useMarketplace } from "./hooks/contracts/useMarketplace";
+export { useNFTCollection } from "./hooks/contracts/useNFTCollection";
+export { usePack } from "./hooks/contracts/usePack";
+export { useToken } from "./hooks/contracts/useToken";
+export { useTokenDrop } from "./hooks/contracts/useTokenDrop";
+export { useVote } from "./hooks/contracts/useVote";
+export { useSplit } from "./hooks/contracts/useSplit";
+export { useMultiwrap } from "./hooks/contracts/useMultiwrap";
+export { useSignatureDrop } from "./hooks/contracts/useSignatureDrop";
 
-export { useStorageUpload, useStorage } from "./hooks/storage";
+export { useStorage } from "./hooks/storage/useStorage";
+export { useStorageUpload } from "./hooks/storage/useStorageUpload";
 export { useNetworkMismatch } from "./hooks/useNetworkMismatch";
 export { useReadonlySDK } from "./hooks/useReadonlySDK";
 export { useSigner } from "./hooks/useSigner";