+++ packages/react-native/src/core/AsyncStorage.ts
@@ -33,7 +33,9 @@ export class AsyncLocalStorage implements IAsyncStorage {
     );
   }
 }
-
+/**
+ * @internal
+ */
 export class LocalStorage implements SyncStorage {
   name: string;
   asyncStorage: MMKV;
@@ -71,6 +73,8 @@ export class noopStorage implements IAsyncStorage {
 }
 
 /**
+ * @internal
+ *
  * Returns a new instance of AsyncLocalStorage
  *
  * @param name - Name to namespace the storage with
@@ -81,6 +85,8 @@ export function createAsyncLocalStorage(name: string) {
 }
 
 /**
+ * @internal
+ *
  * Returns a new instance of LocalStorage
  *
  * @param name - Name to namespace the storage with