+++ packages/react/src/wallet/wallets/safe/types.ts
@@ -1,8 +1,4 @@
 import { WalletConfig } from "@thirdweb-dev/react-core";
 import { SafeWallet } from "@thirdweb-dev/wallets";
 
-export type SafeWalletConfigOptions = {
-  personalWallets?: WalletConfig<any>[];
-};
-
 export type SafeWalletConfig = WalletConfig<SafeWallet>;