+++ packages/wallets/src/evm/connectors/injected/types.ts
@@ -52,6 +52,7 @@ type InjectedProviderFlags = {
   isZerion?: true;
   isOkxWallet?: true;
   isCoreWallet?: true;
+  isOneKey?: true;
   isDefiWallet?: true;
   isRabbyWallet?: true;
   isCoin98Wallet?: true;